// frontend/src/app/core/models/Externos/Residuo.ts
export interface ClasificacionResiduo {
    idClasificacionResiduo: number;
    descripcion: string;
}

export interface TipoResiduo {
    idTipoResiduo: number;
    descripcion: string;
}

export interface UnidadPeso {
    idUnidadMedida: number;
    descripcion: string;
}

export interface FrecuenciaPeso {
    idFrecuenciaResiduo: number;
    descripcion: string;
}